More changes for using uint instead of int for il offsets More changes unifying fault/finally clause types

+++ ILCompiler/TypeSystem/Dnlib/DnlibInstruction.cs
@@ -28,7 +28,7 @@ namespace ILCompiler.TypeSystem.Dnlib
             return ((ILOpcode)code);
         }
 
-        public override int GetSize() => _instruction.GetSize();
+        public override uint GetSize() => (uint)_instruction.GetSize();
 
         public override object Operand
         {
